Update gtfs_headway_analysis.py
Updated to account for frequencies.txt, and use folder of assets with process/data/GTFS (ie. unified data folder with rest of project)   --- that latter change may not be desireable, but the other one is important.  Also includes fixes with taking mean across dates, rather than minimum; and creating a 'direction_id' field, and ensuring this does not have nulls, to avoid if else conditions related to this.