Update gtfs_headway_analysis.py
Updated to output layers for each city with feeds combined; each layer has a new 'feed' attribute to describe the feed from which the stop was sourced.  This means the layers are more directly usable in the global indicators project (subject to filtering by headway).  Also replaced some formatting with f strings to simplify print statements.