Update gtfs_headway_analysis.py
Made validation a configurable parameter, which if False (apparently should be default, as per Basecamp) saves the original stop_times.txt as stop_times_original.txt and reads this in as a Pandas dataframe along with stops.txt, and restricts stops in stop_times.txt to those stops in stops.txt with coordinates within the current bounding box area of interest.  The reason for this is, some feeds (Bern and Odense; both sourced from large national feeds) can have their stop_times departure_time entries largely nullified and otherwise made incorrect resulting in incorrect analysis if this process isn't undertaken; for other feeds, the difference is more subtle.  However, without being clear exactly what in the validation (ostensibly, related to restriction to bounding box, and checking hemisphere) appears to be causing errors, the UrbanAccess gtfsfeed_to_df() functions validation cannot be assumed to be a reliable treatment for input feeds.