Update gtfs_config.py
Updated GTFS configuration with new 'validation':False parameter.  It happens to be set False for all layers, however the main impact is only observable for Odense and Bern.  By setting validation to False, the code now creates a 'cleaned' version of stop_times with stops restricted to those within the limits of the current city's bounding box.