Update _render.sh
I added an extra line in this previously, but just adding one extra (now two) just to make sure that this is resolved - maybe there are differences between editing on windows and running on linux with the line endings impacting the parsing of this.  So, worst case scenario, there's a redundant additional line ending.  Should be good.