Delete jekyll-gh-pages.yml
Remove the jekyll GitHub Actions workflow.  The recommended approach of hosting this on main seems very disruptive from a development perspective (apologies for trying this out); and in retrospect, I think the FAQ should be hosted not within this repository but for the organisation itself (ie. in a separate standalone FAQ repository).  This way, the website can be developed offline and also encompass other topics beyond the spatial indicators software (e.g. 1000 cities challenge aspects and policy indicators)"