Update docker-compose.yml
This commit is as per #256 and allows the example process to be run on a Linux computer by non-root user initialising with `sudo bash ./global-indicators.sh
`, with the following configuration:
- Ubuntu 20.04.4 LTS
- docker-desktop-4.18.0-amd64.deb
- Docker Compose version v2.17.2 (from the docker desktop installation)

While the process ran successfully, there were two apparently unimportant warnings:
```
WARN[0000] The "UID" variable is not set. Defaulting to a blank string. 
WARN[0000] The "GID" variable is not set. Defaulting to a blank string. 
```

Yet, if the 'user' configuration isn't set (line 13) then the request to change permissions using chmod fails with error 'Operation not permitted'

I haven't yet attempted to run this modified docker compose configuration on platforms other than the above.  Ideally we will find a solution that doesn't display a warning...