Create tests.yml
A first go at a minimal test Github Actions workflow, that eventually should support ci/cd as per #258 .   For now, if it works at all (let's see!) it should just run the Docker compose retrieving and launching the image from Dockerhub as a container that is attached to, then running the example analysis workflow.  There are now explicit exit codes used - however, on failure at several points sys.exit is called in the python code returning a printed message that could be interpreted as a non-zero exit code (ie. failure).  So maybe, a successful (ie. without failures) run of the example workflow itself is a kind of holistic test.  Eventually, the idea would be to run more specific tests about specific aspects (and some of these are defined already in the tests folder for use with pre-commit hooks).