Add workflow to sync Wiki to Docs
addresses #569 

Direct monitoring of wiki changes isn't possible, however this script runs a daily check to see if any commits have been made within the last 24 hours.  If so, the wiki will be cloned and copied to the docs folder, which if correctly configured will then trigger a git-hub pages build.